+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.13.12
+
+* Updated the rendering templates of `Section.summary`.
+* Fixed percent calculation in documentation report.
+
 ## 0.13.11+1
 
 * Fixed string processing in untar error path.